# backend/docs/swagger.json
{
    "schemes": [
        "http",
        "https"
    ],
    "swagger": "2.0",
    "info": {
        "description": "AfyaMind telemedicine platform REST API.\n**Authentication:** Most protected endpoints require a JWT Bearer token.\nLogin via POST /auth/login, then click the **Authorize** button above and enter: `Bearer \u003cyour_token\u003e`",
        "title": "AfyaMind API",
        "contact": {
            "name": "AfyaMind Support",
            "email": "support@afyamind.com"
        },
        "license": {
            "name": "MIT",
            "url": "https://opensource.org/licenses/MIT"
        },
        "version": "1.0"
    },
    "host": "localhost:8080",
    "basePath": "/api/v1",
    "paths": {
...
        "/clinics": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Returns a list of all registered healthcare clinics. Super Admin only.",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Clinics"
                ],
                "summary": "List all clinics",
                "responses": {
                    "200": {
                        "description": "List of clinics",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/response.DataEnvelope"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "type": "array",
                                            "items": {
                                                "$ref": "#/definitions/clinics.Clinic"
                                            }
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "401": {
                        "description": "Not authenticated",
                        "schema": {
                            "$ref": "#/definitions/response.ErrorEnvelope"
                        }
                    },
                    "404": {
                        "description": "No clinics found",
                        "schema": {
                            "$ref": "#/definitions/response.ErrorEnvelope"
                        }
                    }
                }
            },
            "post": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Registers a new healthcare clinic and seeds its first admin. Super Admin only.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Clinics"
                ],
                "summary": "Create a new clinic",
                "parameters": [
                    {
                        "description": "Clinic details",
                        "name": "body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/clinics.CreateClinicRequest"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Clinic created successfully",
                        "schema": {
                            "$ref": "#/definitions/clinics.Clinic"
                        }
                    },
                    "400": {
                        "description": "Validation error",
                        "schema": {
                            "$ref": "#/definitions/response.ErrorEnvelope"
                        }
                    },
                    "401": {
                        "description": "Not authenticated",
                        "schema": {
                            "$ref": "#/definitions/response.ErrorEnvelope"
                        }
                    },
                    "409": {
                        "description": "Clinic email already registered or status conflict",
                        "schema": {
                            "$ref": "#/definitions/response.ErrorEnvelope"
                        }
                    }
                }
            }
        },
        "/clinics/{clinicId}": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Retrieves clinical information for a single clinic. Super Admin or Clinic Admin.",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Clinics"
                ],
                "summary": "Get clinic details",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Clinic ID",
                        "name": "clinicId",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Clinic details",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/response.DataEnvelope"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "$ref": "#/definitions/clinics.Clinic"
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "400": {
                        "description": "Validation error",
                        "schema": {
                            "$ref": "#/definitions/response.ErrorEnvelope"
                        }
                    },
                    "401": {
                        "description": "Not authenticated",
                        "schema": {
                            "$ref": "#/definitions/response.ErrorEnvelope"
                        }
                    }
                }
            }
        },
...
        "/clinics/{clinicId}/activate": {
            "patch": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Changes status of the clinic to active. Super Admin only.",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Clinics"
                ],
                "summary": "Activate a clinic",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Clinic ID",
                        "name": "clinicId",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Clinic activated successfully",
                        "schema": {
                            "$ref": "#/definitions/response.MessageEnvelope"
                        }
                    },
                    "400": {
                        "description": "Invalid UUID format",
                        "schema": {
                            "$ref": "#/definitions/response.ErrorEnvelope"
                        }
                    },
                    "401": {
                        "description": "Not authenticated",
                        "schema": {
                            "$ref": "#/definitions/response.ErrorEnvelope"
                        }
                    },
                    "404": {
                        "description": "Clinic not found",
                        "schema": {
                            "$ref": "#/definitions/response.ErrorEnvelope"
                        }
                    }
                }
            }
        },
        "/clinics/{clinicId}/deactivate": {
            "patch": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Changes status of the clinic to deactivated. Super Admin only.",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Clinics"
                ],
                "summary": "Deactivate a clinic",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Clinic ID",
                        "name": "clinicId",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Clinic deactivated successfully",
                        "schema": {
                            "$ref": "#/definitions/response.MessageEnvelope"
                        }
                    },
                    "400": {
                        "description": "Invalid UUID format",
                        "schema": {
                            "$ref": "#/definitions/response.ErrorEnvelope"
                        }
                    },
                    "401": {
                        "description": "Not authenticated",
                        "schema": {
                            "$ref": "#/definitions/response.ErrorEnvelope"
                        }
                    },
                    "404": {
                        "description": "Clinic not found",
                        "schema": {
                            "$ref": "#/definitions/response.ErrorEnvelope"
                        }
                    }
                }
            }
        },
        "/clinics/{clinicId}/doctors": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Lists all doctors registered under a specific clinic. Super Admin or Clinic Admin.",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Clinics"
                ],
                "summary": "List doctors in a clinic",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Clinic ID",
                        "name": "clinicId",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "List of doctors",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/response.DataEnvelope"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "type": "array",
                                            "items": {
                                                "$ref": "#/definitions/clinics.DoctorResponse"
                                            }
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "400": {
                        "description": "Validation error",
                        "schema": {
                            "$ref": "#/definitions/response.ErrorEnvelope"
                        }
                    },
                    "401": {
                        "description": "Not authenticated",
                        "schema": {
                            "$ref": "#/definitions/response.ErrorEnvelope"
                        }
                    }
                }
            }
        },
        "/clinics/{clinicId}/doctors/{doctorId}/activate": {
            "patch": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Activates a doctor's account in the specified clinic. Clinic admin only.",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Clinics"
                ],
                "summary": "Activate a doctor",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Clinic ID",
                        "name": "clinicId",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "Doctor ID",
                        "name": "doctorId",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Doctor activated successfully",
                        "schema": {
                            "$ref": "#/definitions/response.MessageEnvelope"
                        }
                    },
                    "400": {
                        "description": "Validation error",
                        "schema": {
                            "$ref": "#/definitions/response.ErrorEnvelope"
                        }
                    },
                    "401": {
                        "description": "Not authenticated",
                        "schema": {
                            "$ref": "#/definitions/response.ErrorEnvelope"
                        }
                    },
                    "403": {
                        "description": "Forbidden — unauthorized for clinic",
                        "schema": {
                            "$ref": "#/definitions/response.ErrorEnvelope"
                        }
                    },
                    "404": {
                        "description": "Doctor not found",
                        "schema": {
                            "$ref": "#/definitions/response.ErrorEnvelope"
                        }
                    }
                }
            }
        },
        "/clinics/{clinicId}/doctors/{doctorId}/deactivate": {
            "patch": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Deactivates a doctor's account in the specified clinic. Clinic admin only.",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Clinics"
                ],
                "summary": "Deactivate a doctor",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Clinic ID",
                        "name": "clinicId",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "Doctor ID",
                        "name": "doctorId",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Doctor deactivated successfully",
                        "schema": {
                            "$ref": "#/definitions/response.MessageEnvelope"
                        }
                    },
                    "400": {
                        "description": "Validation error",
                        "schema": {
                            "$ref": "#/definitions/response.ErrorEnvelope"
                        }
                    },
                    "401": {
                        "description": "Not authenticated",
                        "schema": {
                            "$ref": "#/definitions/response.ErrorEnvelope"
                        }
                    },
                    "403": {
                        "description": "Forbidden — unauthorized for clinic",
                        "schema": {
                            "$ref": "#/definitions/response.ErrorEnvelope"
                        }
                    },
                    "404": {
                        "description": "Doctor not found",
                        "schema": {
                            "$ref": "#/definitions/response.ErrorEnvelope"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "clinics.Clinic": {
            "type": "object",
            "properties": {
                "address": {
                    "type": "string"
                },
                "created_at": {
                    "type": "string"
                },
                "email": {
                    "type": "string"
                },
                "id": {
                    "type": "string"
                },
                "name": {
                    "type": "string"
                },
                "phone": {
                    "type": "string"
                },
                "status": {
                    "type": "string"
                },
                "updated_at": {
                    "type": "string"
                }
            }
        },
        "clinics.CreateClinicRequest": {
            "type": "object",
            "required": [
                "admin_first_name",
                "admin_last_name",
                "email",
                "name",
                "phone"
            ],
            "properties": {
                "address": {
                    "description": "Street address of the clinic",
                    "type": "string",
                    "example": "123 Medical Plaza, Nairobi"
                },
                "admin_first_name": {
                    "description": "First name of the designated clinic administrator",
                    "type": "string",
                    "example": "Alice"
                },
                "admin_last_name": {
                    "description": "Last name of the designated clinic administrator",
                    "type": "string",
                    "example": "Smith"
                },
                "email": {
                    "description": "Main contact email address",
                    "type": "string",
                    "example": "contact@stlukes.com"
                },
                "name": {
                    "description": "Name of the clinic",
                    "type": "string",
                    "example": "St. Luke Health Center"
                },
                "phone": {
                    "description": "Telephone contact number",
                    "type": "string",
                    "example": "+254711223344"
                }
            }
        },
        "clinics.DoctorResponse": {
            "type": "object",
            "properties": {
                "created_at": {
                    "type": "string"
                },
                "doctor_status": {
                    "type": "string",
                    "example": "active"
                },
                "email": {
                    "type": "string",
                    "example": "jane.doe@example.com"
                },
                "first_name": {
                    "type": "string",
                    "example": "Jane"
                },
                "id": {
                    "type": "string",
                    "example": "550e8400-e29b-41d4-a716-446655440002"
                },
                "invited_by": {
                    "type": "string",
                    "example": "550e8400-e29b-41d4-a716-446655440001"
                },
                "last_name": {
                    "type": "string",
                    "example": "Doe"
                },
                "license_number": {
                    "type": "string",
                    "example": "LIC-12345"
                },
                "phone": {
                    "type": "string",
                    "example": "+254712345678"
                },
                "role": {
                    "type": "string",
                    "example": "doctor"
                },
                "specialization": {
                    "type": "string",
                    "example": "Pediatrics"
                }
            }
        },
...
        "response.DataEnvelope": {
            "description": "Standard API success response",
            "type": "object",
            "properties": {
                "data": {
                    "description": "The response payload - structure varies per endpoint"
                }
            }
        },
        "response.ErrorBody": {
            "type": "object",
            "properties": {
                "code": {
                    "type": "string"
                },
                "message": {
                    "type": "string"
                }
            }
        },
        "response.ErrorEnvelope": {
            "type": "object",
            "properties": {
                "error": {
                    "$ref": "#/definitions/response.ErrorBody"
                }
            }
        },
        "response.MessageBody": {
            "type": "object",
            "properties": {
                "message": {
                    "type": "string",
                    "example": "Operation completed successfully"
                }
            }
        },
        "response.MessageEnvelope": {
            "description": "Simple message response",
            "type": "object",
            "properties": {
                "data": {
                    "description": "Human-readable status message",
                    "allOf": [
                        {
                            "$ref": "#/definitions/response.MessageBody"
                        }
                    ]
                }
            }
        },
...
    },
    "securityDefinitions": {
        "BearerAuth": {
            "description": "JWT access token. Format: **Bearer \u0026lt;token\u0026gt;**",
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    }
}